LevelManager Split - building lib-tempo - Need to implement a few getters and setters
