+++ src/lib-tempo/CMakeLists.txt
@@ -7,6 +7,8 @@ set(TEMPO_LIB_TEMPO_HEADERS
 	"${CMAKE_CURRENT_LIST_DIR}/include/tempo/time.hpp"
 	"${CMAKE_CURRENT_LIST_DIR}/include/tempo/Types.hpp"
 	"${CMAKE_CURRENT_LIST_DIR}/include/tempo/component/ComponentGridAi.hpp"
+	"${CMAKE_CURRENT_LIST_DIR}/include/tempo/entity/ComponentGridMotion.hpp"
+	"${CMAKE_CURRENT_LIST_DIR}/include/tempo/entity/ComponentGridPosition.hpp"
 	"${CMAKE_CURRENT_LIST_DIR}/include/tempo/component/ComponentHealth.hpp"
 	"${CMAKE_CURRENT_LIST_DIR}/include/tempo/component/ComponentRender.hpp"
 	"${CMAKE_CURRENT_LIST_DIR}/include/tempo/component/ComponentTransform.hpp"
@@ -15,7 +17,6 @@ set(TEMPO_LIB_TEMPO_HEADERS
 	"${CMAKE_CURRENT_LIST_DIR}/include/tempo/entity/EntityCreationServer.hpp"
 	"${CMAKE_CURRENT_LIST_DIR}/include/tempo/entity/ID.hpp"
 	"${CMAKE_CURRENT_LIST_DIR}/include/tempo/entity/RenderHealth.hpp"
-	"${CMAKE_CURRENT_LIST_DIR}/include/tempo/entity/LevelManager.hpp"
 	"${CMAKE_CURRENT_LIST_DIR}/include/tempo/entity/LevelRenderer.hpp"
 	"${CMAKE_CURRENT_LIST_DIR}/include/tempo/entity/PlayerLocal.hpp"
 	"${CMAKE_CURRENT_LIST_DIR}/include/tempo/entity/PlayerRemote.hpp"
@@ -27,6 +28,7 @@ set(TEMPO_LIB_TEMPO_HEADERS
 	"${CMAKE_CURRENT_LIST_DIR}/include/tempo/network/server.hpp"
 	"${CMAKE_CURRENT_LIST_DIR}/include/tempo/system/SystemGridAi.hpp"
 	"${CMAKE_CURRENT_LIST_DIR}/include/tempo/system/SystemHealth.hpp"
+	"${CMAKE_CURRENT_LIST_DIR}/include/tempo/entity/SystemLevelManager.hpp"
 	"${CMAKE_CURRENT_LIST_DIR}/include/tempo/system/SystemRender.hpp"
 	"${CMAKE_CURRENT_LIST_DIR}/include/tempo/system/SystemTransform.hpp"
 )
@@ -37,6 +39,8 @@ set(TEMPO_LIB_TEMPO_SOURCE
 	"${CMAKE_CURRENT_LIST_DIR}/src/song.cpp"
 	"${CMAKE_CURRENT_LIST_DIR}/src/structures.cpp"
 	"${CMAKE_CURRENT_LIST_DIR}/src/time.cpp"
+	"${CMAKE_CURRENT_LIST_DIR}/src/entity/ComponentGridMotion.cpp"
+	"${CMAKE_CURRENT_LIST_DIR}/src/entity/ComponentGridPosition.cpp"
 	"${CMAKE_CURRENT_LIST_DIR}/src/component/ComponentHealth.cpp"
 	"${CMAKE_CURRENT_LIST_DIR}/src/component/ComponentRender.cpp"
 	"${CMAKE_CURRENT_LIST_DIR}/src/component/ComponentTransform.cpp"
@@ -45,7 +49,6 @@ set(TEMPO_LIB_TEMPO_SOURCE
 	"${CMAKE_CURRENT_LIST_DIR}/src/entity/EntityCreationServer.cpp"
 	"${CMAKE_CURRENT_LIST_DIR}/src/entity/ID.cpp"
 	"${CMAKE_CURRENT_LIST_DIR}/src/entity/RenderHealth.cpp"
-	"${CMAKE_CURRENT_LIST_DIR}/src/entity/LevelManager.cpp"
 	"${CMAKE_CURRENT_LIST_DIR}/src/entity/LevelRenderer.cpp"
 	"${CMAKE_CURRENT_LIST_DIR}/src/entity/PlayerLocal.cpp"
 	"${CMAKE_CURRENT_LIST_DIR}/src/entity/PlayerRemote.cpp"
@@ -57,6 +60,7 @@ set(TEMPO_LIB_TEMPO_SOURCE
 	"${CMAKE_CURRENT_LIST_DIR}/src/network/server.cpp"
 	"${CMAKE_CURRENT_LIST_DIR}/src/system/SystemGridAi.cpp"
 	"${CMAKE_CURRENT_LIST_DIR}/src/system/SystemHealth.cpp"
+	"${CMAKE_CURRENT_LIST_DIR}/src/entity/SystemLevelManager.cpp"
 	"${CMAKE_CURRENT_LIST_DIR}/src/system/SystemRender.cpp"
 	"${CMAKE_CURRENT_LIST_DIR}/src/system/SystemTransform.cpp"
 )
