+++ src/lib-tempo/include/tempo/component/ComponentTransform.hpp
@@ -4,7 +4,8 @@
 #include <anax/System.hpp>
 #include <anax/Component.hpp>
 
-#include <tempo/entity/LevelManager.hpp>
+#include <tempo/entity/ComponentGridMotion.hpp>
+#include <tempo/entity/ComponentGridPosition.hpp>
 
 #include <OgreMath.h>
 #include <OgreVector3.h>
