+++ src/lib-tempo/include/tempo/entity/EntityCreationClient.hpp
@@ -10,9 +10,9 @@
 #include <tempo/component/ComponentRender.hpp>
 #include <tempo/entity/EntityCreation.hpp>
 #include <tempo/entity/ID.hpp>
-#include <tempo/entity/LevelManager.hpp>
 #include <tempo/entity/PlayerLocal.hpp>
 #include <tempo/entity/PlayerRemote.hpp>
+#include <tempo/entity/SystemLevelManager.hpp>
 
 namespace tempo
 {
