+++ src/lib-tempo/include/tempo/entity/EntityCreationServer.hpp
@@ -6,8 +6,8 @@
 #include <tempo/component/ComponentGridAi.hpp>
 #include <tempo/entity/EntityCreation.hpp>
 #include <tempo/entity/ID.hpp>
-#include <tempo/entity/LevelManager.hpp>
 #include <tempo/entity/PlayerRemoteServer.hpp>
+#include <tempo/entity/SystemLevelManager.hpp>
 
 #include <string.h>
