+++ src/lib-tempo/include/tempo/system/SystemGridAi.hpp
@@ -4,7 +4,7 @@
 #include <anax/Component.hpp>
 
 #include <tempo/component/ComponentGridAi.hpp>
-#include <tempo/entity/LevelManager.hpp>
+#include <tempo/entity/ComponentGridMotion.hpp>
 
 namespace tempo{
 	struct SystemGridAi : anax::System<anax::Requires<ComponentGridMotion, ComponentGridAi>> {
