+++ src/lib-tempo/include/tempo/system/SystemTransform.hpp
@@ -4,7 +4,9 @@
 #include <anax/System.hpp>
 #include <anax/Component.hpp>
 
-#include <tempo/entity/LevelManager.hpp>
+#include <tempo/entity/ComponentGridMotion.hpp>
+#include <tempo/entity/ComponentGridPosition.hpp>
+#include <tempo/entity/SystemLevelManager.hpp>
 #include <tempo/component/ComponentTransform.hpp>
 
 #include <OgreMath.h>
